requirements.txt
allure-pytest==2.9.45
pytest==7.1.2
selenium==4.16.0